refactor: split agent studio header
